[diffusion] chore: minor code cleanups
