+++ python/sglang/multimodal_gen/runtime/models/registry.py
@@ -316,8 +316,9 @@ class _ModelRegistry:
         normalized_arch = []
         for arch in architectures:
             if arch not in self.registered_models:
+                registered_models = list(self.registered_models.keys())
                 raise Exception(
-                    f"Unsupported model architecture: {arch}. Registered architectures: {self.registered_models=}"
+                    f"Unsupported model architecture: {arch}. Registered architectures: {registered_models}"
                 )
             normalized_arch.append(arch)
         return normalized_arch
